% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ACEFScore.R
\name{ACEFScore}
\alias{ACEFScore}
\title{Compute ACEF Score}
\usage{
ACEFScore(df = data.frame())
}
\arguments{
\item{df}{data frame, must contain age, ejectionFraction as rows}
}
\description{
Compute the ACEF Score by the formula from XY
}

